Split Appointmnet page (AppointmentForm component and Appointment page)
