+++ src/components/AppointmentForm/style.module.css
@@ -1,0 +1,59 @@
+.form {
+  color: #fff;
+  margin-top: 1rem;
+  display: grid;
+  grid-template-columns: repeat(3, 1fr);
+  gap: 1.5rem;
+}
+
+.form-child,
+.btn {
+  border: 1px solid #fff;
+  border-radius: 2rem;
+  padding: 1rem;
+  text-align: center;
+}
+
+.form-child {
+  background-color: rgba(151, 191, 15, 0.1);
+}
+
+.btn {
+  background-color: #fff;
+  color: var(--green);
+  grid-column: 1 / span 3;
+  width: 200px;
+  margin: auto;
+}
+
+::placeholder {
+  color: #fff;
+  opacity: 1;
+}
+
+::-ms-input-placeholder {
+  color: #fff;
+}
+
+:-ms-input-placeholder {
+  color: #fff;
+}
+
+.errors {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  color: #fff;
+}
+
+@media only screen and (max-width: 670px) {
+  .form {
+    grid-template-columns: 1fr;
+  }
+
+  .btn {
+    grid-column: unset;
+    width: unset;
+    margin: unset;
+  }
+}
